Add minor format and move components to folder
