+++ src/App.tsx
@@ -1,8 +1,8 @@
-import Archive from "./Archive";
-import Footer from "./Footer";
-import IntroSection from "./IntroSection";
-import LatestSection from "./LatestSection";
-import MenuBar from "./MenuBar";
+import Archive from "./components/Archive";
+import Footer from "./components/Footer";
+import IntroSection from "./components/IntroSection";
+import LatestSection from "./components/LatestSection";
+import MenuBar from "./components/MenuBar";
 import './styles.css'
 
 function App(): JSX.Element {
