+++ src/components/LatestSection.tsx
