+++ src/components/MenuBar.tsx
@@ -1,14 +1,12 @@
 export default function MenuBar():JSX.Element {
     return (
         <>
-            <p>menu section START</p>
             <ul id='menu_list'>
                 <li>Niamh Brockbank</li>
                 <li>Featured</li>
                 <li>Archive</li>
                 <li>Info</li>
             </ul>
-            <p>menu section END</p>
             <hr />
         </>
     )
