+++ src/styles.css
@@ -1,5 +1,13 @@
+/*
+colour palette
+#686963 dim gray
+#f9f9f5 (off-white)
+*/
+
 body{
-    color : red
+    font-family: 'Gill Sans', 'Gill Sans MT', Calibri, 'Trebuchet MS', sans-serif, sans-serif;
+    color : #41423e;
+    background-color: #f9f9f5;
 }
 
 #menu_list{
